Everything works - except the pictues
The new file server doesn't decode the path; Therefore it doesn't work with anything other than english letters and basically no special characters like spaces
