+++ routes/index.tsx
@@ -8,14 +8,13 @@ import {
 } from "@/components/home/mod.tsx";
 import type { PageProps } from "fresh";
 import type { State } from "@/core/types.ts";
-import { Head } from "$fresh/runtime.ts";
 
 export default function Home(props: PageProps<never, State>) {
 	const language = props.state.language;
 
 	return (
 		<>
-			<Head>
+			<head>
 				<meta property="og:title" content="Felix Schindler" />
 				<meta property="og:url" content="https://www.schindlerfelix.de/" />
 				<meta
@@ -23,7 +22,7 @@ export default function Home(props: PageProps<never, State>) {
 					content="https://www.schindlerfelix.de/img/icons/icon-512x512.png"
 				/>
 				<meta property="og:type" content="website" />
-			</Head>
+			</head>
 			<div class="flex flex-col gap-5">
 				<Intro lang={language} />
 				<About lang={language} />
